collision system arreglado + enemigos recuperan velocidad al acabarse el terremoto

+++ ProyectoH/src/systems/EnemySystem.cpp
@@ -58,7 +58,7 @@ void  EnemySystem::receive(const Message& m) {
 		}
 		break;
 	case _m_ENTITY_TO_ATTACK:
-		if (m.entity_to_attack.e != nullptr && mngr_->isAlive(m.entity_to_attack.e)) {
+		if (m.entity_to_attack.targetId == _hdlr_ENEMIES && m.entity_to_attack.e != nullptr && mngr_->isAlive(m.entity_to_attack.e)) {
 			HealthComponent* h = mngr_->getComponent<HealthComponent>(m.entity_to_attack.e);
 			if (h->subtractHealth(m.entity_to_attack.damage)) {
 				mngr_->deleteHandler(_hdlr_ENEMIES, m.entity_to_attack.e);
